added conditional text for licence confirmation method screen
