+++ packages/gafl-webapp-service/src/pages/contact/digital-licence/licence-confirmation-method/route.js
@@ -12,6 +12,7 @@ import { mobilePhoneValidator } from '../../../../processors/contact-validator.j
 export const getData = async request => {
   const { change } = request?.query
   const permission = await request.cache().helpers.transaction.getCurrentPermission()
+  const { isLicenceForYou } = await request.cache().helpers.status.getCurrentPermission()
 
   // page is only permitted for physical licences
   if (!isPhysical(permission)) {
@@ -19,6 +20,7 @@ export const getData = async request => {
   }
 
   return {
+    isLicenceForYou,
     licensee: permission.licensee,
     howContacted: HOW_CONTACTED,
     ...(change === 'email' && { changeEmail: true }),
